ftbuffer works with basic header

+++ applications/mne_scan/plugins/ftbuffer/ftbuffproducer.h
@@ -153,7 +153,7 @@ signals:
     void extendedHeaderChunks(SimpleStorage chunkData);
 
 
-    void bufferParameters(QPair<int,int> numChanandFreq);
+    void bufferParameters(QPair<int,float> numChanandFreq);
 
 private:
